+++ src/main/java/ar/edu/unlam/nuralign/application/ports/in/patientTherapist/DeletePatientTherapistAssignationUseCase.java
@@ -1,5 +1,5 @@
 package ar.edu.unlam.nuralign.application.ports.in.patientTherapist;
 
 public interface DeletePatientTherapistAssignationUseCase {
-    void deletePatientTherapistAssignation(Long patientTherapistId);
+    void deletePatientTherapistAssignation(Long patientId, Long therapistId);
 }